requirements.txt
Flask==2.3.3
mysql-connector-python==8.3.0
pygame==2.5.2
pymunk==6.6.0
